fix erroneously setting license_id of variant frames to their own id
Refactor the computed property of availableVariants (used to display the list of potential "parent" frames in the dropdown) to use a newly created availableFrames property. The availableFrames property does the same job as the previous availableVariants property, but exposes just all frames instead of only their names. The avilableVariants property now merely reads the names out of the availableFrames property.

The reason for the refactor is to allow for automatically matching `variant` fields to the corresponding `license_id` for the "parent" frame. This is just a search across the availableFrames property to find which frame has the right name. This assumes that the `variant` field will always case-sensitive match the `name` of the "parent" frame.