Allow users to make variant frames for Long Rim/Wallflower/KTB licenses
The LCP editor has no knowledge of frames it can't source directly from @massif/lancer-data (reasonable enough). Unfortunately, this means that variant frames can't auto-infer the correct license_id to use if the license they're trying to spin off isn't in that namespace. To counteract this, hardcode a list of the currently existing new licenses added in various supplements (the current pile is Long Rim, Wallflower, KTB, & OWS). This is not necessary for something like SSMR, for example, since only new _licenses_ need to be added here.

In a better world, I coded this so that the LCP editor allows you to just specify a license_id separately, but this is probably the best solution in terms of "making it simple for users to use without in-depth knowledge of LCP rules".