Fix Core Bonuses behaving oddly with adding multiple synergies
It was not resetting the "isEdit" state when attempting to create a new item, leading to weird things where it'd assume you're still wanting to overwrite the previous synergy. Now, this shouldn't happen for core bonuses specifically, and it shouldn't happen when messing with synergies anywhere (exiting any synergy pane will always mark you as not editing, and you'll only be set to editing again if you actually try to edit something directly). I'm not 100% sure that doesn't have weird effects, but I've tested it on systems, core bonuses and weapons and it seems to function OK.