Allow for using fancy Horus text in the LCP editor :)
Adds two little icons to the text editor, so they're embedded in every text entry instance in the LCP editor. The sawtooth wave is for HORUS codeblock formatting, and the sine wave is for HORUS 'subtle' text. No promises that the buttons are very intuitive, but they _should_ work. The only finicky part is really that neither of the buttons are toggles, they only _apply_ the formatting, so you'd need to delete them (or in the case of horus-code formatting, just toggle it away from code to clear the formatting) to reset it.