Fix (?) banner preview for frame data
I'm honestly not sure wtf is up with the old code. It's just copy-pasted from CompCon's codebase, which is presumably working since it works fine on there - however, I did notice that CompCon's compiled _output_ css is different. The code is an image inside a div inside a div, and the LCP editor turns it into an image + a div together, inside a div. I ended up just copying what CompCon's version turns into, which uses background-image on the div for the banner, and that works much better. The x-position is a little funky but that's not relevant for the preview purposes, I believe.