Begin the process of correcting NPC imports
NPC classes now import correctly, and display features as hoped. NPC templates import but I haven't gotten around to working on that pane so they don't display. Export is untouched, and editing does not work yet.

Also, fixed the warning I was ignoring about double-including the TipTap Code extension (since it was in the Starter Kit & also imported separately for my code override). I've just adjusted this to not use the starter kit and instead depend on the individual extensions. Go figure.