add dependencies selection
bad bad bad solution. i'm pushing this as a temporary stand-in so that it's at least functional, but I'm rapidly approaching the limits of my vue abilities. reactive arrays seem to be blanket illegal, even when handled property (i.e. via v-model / a passed reference?).

specifically, the LOAD_LCP hook that keeps firing is running into "ruh roh! I hate trying to call .join() on this array property, because its child components are not strings". I genuinely cannot tell if this is a reactivity issue in general (backtrace implies as much) and I just don't know how to do that, or if it's a weirdness about how LOAD_LCP is handled.