revamp ID editing
the downsides:
- more annoying to edit the ID (you have to do a pop-up menu, you can't just edit it via the text box anymore)

the upsides:
- now, you can actually change the ID of an item without it duplicating the item, requiring you to delete the old one.
- in addition, changing the ID of a manufacturer will actually propagate to all of the items correclty

does NOT work for frames with licensed items. I did not want to mess with that box of worms, since it would involve doing a full-find/replace identical to the main page replace ID tool. for frames with child items detected, it will tell you to just use that button instead (so that the only destructive-behavior button is there)